+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## [53.7.4]
+- Set semantic description on segmented control items
+
+## [53.7.4]
 - Fix error causing index out of range in `TabBadgeService` when tabs have changed 
 
 ## [53.7.3]